Add function to compile the model via onnx-mlir command
