+++ include/OnnxMlirCompiler.h
@@ -111,6 +111,27 @@ ONNX_MLIR_EXPORT void omClearCompilerOption(const OptionKind kind);
 ONNX_MLIR_EXPORT const char *omGetCompilerOption(const OptionKind kind);
 
 /*!
+ *  C interface to compile an onnx model from a file via onnx-mlir command.
+ *
+ *  @param inputFilename File name pointing onnx model protobuf or MLIR.
+ *  Name may include a path, and must include the file name and its extention.
+ *  @param outputBaseName File name without extension to write output.
+ *  Name may include a path, must include the file name, and should not include
+ * an extention.
+ *  @param emissionTarget Target format to compile to.
+ *  @param flags A char * contains all the options provided to compile the
+ * model.
+ *  @param outputFilename Output file name of the compiled output for the given
+ * emission target. User is responsible for freeing the string.
+ *  @param errorMessage Output error message, if any. User is responsible for
+ * freeing the string.
+ *  @return 0 on success or OnnxMlirCompilerErrorCodes on failure.
+ */
+ONNX_MLIR_EXPORT int64_t omCompileFromFileViaCommand(const char *inputFilename,
+    const char *outputBaseName, EmissionTargetType emissionTarget,
+    const char *flags, const char **outputFilename, const char **errorMessage);
+
+/*!
  *  Compile an onnx model from a file containing MLIR or ONNX protobuf. When
  *  generating libraries or jar files, the compiler will link in lightweight
  *  runtimes / jar files. If these libraries / jar files are not in the system
