+++ src/Compiler/OnnxMlirCompiler.cpp
@@ -4,6 +4,10 @@
 
 #include "OnnxMlirCompiler.h"
 #include "CompilerUtils.hpp"
+#include "ExternalUtil.hpp.in"
+
+using namespace mlir;
+using namespace onnx_mlir;
 
 extern "C" {
 namespace onnx_mlir {
@@ -47,6 +51,33 @@ ONNX_MLIR_EXPORT void omClearCompilerOption(const OptionKind kind) {
 ONNX_MLIR_EXPORT const char *omGetCompilerOption(const OptionKind kind) {
   std::string val = getCompilerOption(kind);
   return strdup(val.c_str());
+}
+
+ONNX_MLIR_EXPORT int64_t omCompileFromFileViaCommand(const char *inputFilename,
+    const char *outputBaseName, EmissionTargetType emissionTarget,
+    const char *flags, const char **outputFilename, const char **errorMessage) {
+  // Manually process the flags
+  // Save the result string vector after processing
+  std::vector<std::string> flagsVector;
+  // Use the same standard as std::isspace to define white space characters
+  const char delimiters[6] = {0x20, 0x0c, 0x0a, 0x0d, 0x09, 0x0b};
+  // Use strtok_r instead of strtok because strtok_r is thread safe
+  char *token;
+  char *buffer = new char[std::strlen(flags) + 1];
+  std::strcpy(buffer, flags);
+  char *rest = buffer;
+  while ((token = strtok_r(rest, delimiters, &rest)) != NULL) {
+    flagsVector.push_back(std::string(token));
+  }
+  // Use 'onnx-mlir' command to compile the model.
+  std::string onnxmlirPath = getToolPath("onnx-mlir");
+  struct Command onnxmlirCompile(
+      /*exePath=*/!onnxmlirPath.empty() ? onnxmlirPath : kOnnxmlirPath);
+  for (std::size_t i = 0; i < flagsVector.size(); i++) {
+    onnxmlirCompile.appendStr(flagsVector[i]);
+  }
+  int rc = onnxmlirCompile.exec();
+  return rc != 0 ? CompilerFailureInLLVMOpt : CompilerSuccess;
 }
 
 ONNX_MLIR_EXPORT int64_t omCompileFromFile(const char *inputFilename,
